refactor: calculate rebalance amount+fees offchain for tests
